LG-4605: move the mock proofing classes into their own module
