+++ app/jobs/resolution_proofing_job.rb
@@ -189,7 +189,7 @@ class ResolutionProofingJob < ApplicationJob
 
   def resolution_proofer
     @resolution_proofer ||= if IdentityConfig.store.proofer_mock_fallback
-      Proofing::ResolutionMockClient.new
+      Proofing::Mock::ResolutionMockClient.new
     else
       LexisNexis::InstantVerify::Proofer.new(
         instant_verify_workflow: IdentityConfig.store.lexisnexis_instant_verify_workflow,
@@ -205,7 +205,7 @@ class ResolutionProofingJob < ApplicationJob
 
   def state_id_proofer
     @state_id_proofer ||= if IdentityConfig.store.proofer_mock_fallback
-      Proofing::StateIdMockClient.new
+      Proofing::Mock::StateIdMockClient.new
     else
       Aamva::Proofer.new(
         auth_request_timeout: IdentityConfig.store.aamva_auth_request_timeout,
