+++ spec/features/sp_cost_tracking_spec.rb
@@ -29,11 +29,11 @@ feature 'SP Costing', :email do
     expect_sp_cost_type(3, 2, 'acuant_result')
     expect_sp_cost_type(
       4, 2, 'lexis_nexis_resolution',
-      transaction_id: Proofing::ResolutionMockClient::TRANSACTION_ID
+      transaction_id: Proofing::Mock::ResolutionMockClient::TRANSACTION_ID
     )
     expect_sp_cost_type(
       5, 2, 'aamva',
-      transaction_id: Proofing::StateIdMockClient::TRANSACTION_ID
+      transaction_id: Proofing::Mock::StateIdMockClient::TRANSACTION_ID
     )
     expect_sp_cost_type(6, 2, 'lexis_nexis_address')
     expect_sp_cost_type(7, 2, 'user_added')
